+++ cmd/info.go
@@ -9,7 +9,10 @@ import (
 
 	cmdutil "github.com/tranvictor/jarvis/cmd/util"
 	"github.com/tranvictor/jarvis/config"
+	jarvisnetworks "github.com/tranvictor/jarvis/networks"
+	"github.com/tranvictor/jarvis/txanalyzer"
 	"github.com/tranvictor/jarvis/util"
+	"github.com/tranvictor/jarvis/util/reader"
 )
 
 var txCmd = &cobra.Command{
@@ -24,7 +27,7 @@ var txCmd = &cobra.Command{
 		tc, _ := cmdutil.TxContextFrom(cmd)
 
 		para := strings.Join(args, " ")
-		txs := util.ScanForTxHashes(para)
+		nwks, txs := cmdutil.ScanForTxs(para)
 		if len(txs) == 0 {
 			appUI.Error("Couldn't find any tx hash in the params")
 			return
@@ -33,7 +36,7 @@ var txCmd = &cobra.Command{
 		if len(txs) > 1 {
 			appUI.Info("Analyzing %d transactions:", len(txs))
 			for i, t := range txs {
-				appUI.Info("  %d. %s", i+1, t)
+				appUI.Info("  %d. %s:%s", i+1, nwks[i], t)
 			}
 		}
 
@@ -48,29 +51,71 @@ var txCmd = &cobra.Command{
 			}()
 		}
 
+		envs := map[string]infoTxEnv{}
 		for i, t := range txs {
 			if i > 0 {
 				appUI.Info("")
 			}
 			if len(txs) > 1 {
-				appUI.Info("%s", t)
+				appUI.Info("%s:%s", nwks[i], t)
+			}
+			env, err := infoTxEnvFor(tc, envs, nwks[i])
+			if err != nil {
+				appUI.Error("%s network is not supported. Skip.", nwks[i])
+				continue
 			}
 			d := util.AnalyzeAndPrint(
 				appUI,
-				tc.Reader,
-				tc.Analyzer,
+				env.reader,
+				env.analyzer,
 				t,
-				config.Network(),
+				env.network,
 				config.ForceERC20ABI,
 				config.CustomABI,
 				nil,
 				nil,
 				util.InfoLayout(config.DegenMode),
-				cmdutil.InfoClearSign(config.Network()),
+				cmdutil.InfoClearSign(env.network),
 			)
 			displays[t] = d
 		}
 	},
+}
+
+type infoTxEnv struct {
+	network  jarvisnetworks.Network
+	reader   reader.Reader
+	analyzer util.TxAnalyzer
+}
+
+func infoTxEnvFor(tc cmdutil.TxContext, cache map[string]infoTxEnv, name string) (infoTxEnv, error) {
+	if env, ok := cache[name]; ok {
+		return env, nil
+	}
+	n, err := jarvisnetworks.GetNetwork(name)
+	if err != nil {
+		return infoTxEnv{}, err
+	}
+	key := n.GetName()
+	if env, ok := cache[key]; ok {
+		cache[name] = env
+		return env, nil
+	}
+	env := infoTxEnv{network: n}
+	if tc.Reader != nil && key == config.Network().GetName() {
+		env.reader = tc.Reader
+		env.analyzer = tc.Analyzer
+	} else {
+		r, err := util.EthReader(n)
+		if err != nil {
+			return infoTxEnv{}, err
+		}
+		env.reader = r
+		env.analyzer = txanalyzer.NewGenericAnalyzer(r, n)
+	}
+	cache[name] = env
+	cache[key] = env
+	return env, nil
 }
 
 func init() {